requirements.txt
appdirs==1.4.4
asgiref==3.2.10
boto3==1.14.32
celery==5.0.0
Cython==0.29.23
decouple==0.0.7
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-celery-beat==2.0.0
django-ckeditor==6.0.0
django-countries==7.2.1
django-crispy-forms==1.9.2
django-friendship==1.8.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-import-export==2.5.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-storages==1.9.1
django-taggit==1.3.0
django-timezone-field==4.1.2
docutils==0.15.2
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gunicorn==20.0.4
Pillow==7.1.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
setuptools-git==1.2
sqlparse==0.3.1
tablib==3.0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.60.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.35
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.3.1
